moving settings inside the package
